move comparers to own module
