+++ mitxgraders/comparers/__init__.py
@@ -1,0 +1,5 @@
+from comparers import eigenvector_comparer
+
+__all__ = [
+    'eigenvector_comparer'
+]
