+++ mitxgraders/formulagrader/matrixgrader.py
@@ -5,7 +5,6 @@ Defines a FormulaGrader subtype that handles matrices, too.
 """
 from numbers import Number
 from collections import namedtuple
-import numpy as np
 from mitxgraders.formulagrader.formulagrader import FormulaGrader
 from voluptuous import Required, Any
 from mitxgraders.helpers.validatorfuncs import NonNegative
@@ -14,7 +13,6 @@ from mitxgraders.helpers.calc.exceptions import (
     CalcError, MathArrayShapeError as ShapeError)
 from mitxgraders.helpers.calc.mathfuncs import (
     merge_dicts, ARRAY_ONLY_FUNCTIONS)
-from mitxgraders.helpers.calc.formatters import get_description
 
 class InputTypeError(CalcError):
     """
@@ -175,28 +173,3 @@ class MatrixGrader(FormulaGrader):
         utils.validate_shape(student_input, shape)
 
         return utils.within_tolerance(expected_input, student_input)
-
-def eigenvector_comparer(comparer_params, student_input, utils):
-    """
-    A comparer function for MatrixGrader.
-
-    Example Usage:
-
-    """
-
-    matrix, eigenvalue = comparer_params
-
-    # matrix is square with shape (n, n); student input should have shape (n, )
-    expected_input_shape = (matrix.shape[0], )
-    utils.validate_shape(student_input, expected_input_shape)
-
-    diff = matrix*student_input - eigenvalue * student_input
-    return utils.within_tolerance(0, np.linalg.norm(diff))
-
-
-
-
-
-
-
-#
